% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bearer_auth.r
\name{is_bearer_auth}
\alias{is_bearer_auth}
\title{Get Bearer API Authentication token}
\usage{
is_bearer_auth(secret_key, refresh_token)
}
\arguments{
\item{secret_key}{Account Secret key}

\item{refresh_token}{Account Refresh token}
}
\description{
Get Bearer API Authentication token <https://developers.is.com/ironsource-mobile/air/authentication>
}
