% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_load_data.r
\name{is_load_data}
\alias{is_load_data}
\title{Load IronSource Data from Link}
\usage{
is_load_data(link, file_path)
}
\arguments{
\item{link}{Link gathered from IronSource API}

\item{file_path}{Destination path including the file name}
}
\description{
Load IronSource Data from Link
}
