% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_reporting.r
\name{is_reporting}
\alias{is_reporting}
\title{Get Data from Reporting API}
\usage{
is_reporting(
  start_date,
  end_date,
  bearer_token,
  app_key = NULL,
  country = NULL,
  ad_units = NULL,
  ad_source = NULL,
  metrics = NULL,
  breakdowns = NULL
)
}
\arguments{
\item{start_date}{Start date of report in UTC timezone}

\item{end_date}{End date of report in UTC timezone}

\item{bearer_token}{Bearer API Authentication token}

\item{app_key}{Application key filter}

\item{country}{Country filter}

\item{ad_units}{Ad units filter}

\item{ad_source}{Ad source filter}

\item{metrics}{Metrics filter}

\item{breakdowns}{Breakdowns for the report}
}
\description{
Get Data from Reporting API <https://developers.is.com/ironsource-mobile/air/reporting/>
}
