% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_revenue_impression.r
\name{is_revenue_impression}
\alias{is_revenue_impression}
\title{Get Impression Level Revenue Data from Ad Revenue Measurement API}
\usage{
is_revenue_impression(date, app_key, bearer_token)
}
\arguments{
\item{date}{Date of report in UTC timezone}

\item{app_key}{Application key}

\item{bearer_token}{Bearer API Authentication token}
}
\description{
Get Impression Level Revenue Data from Ad Revenue Measurement API <https://developers.is.com/ironsource-mobile/air/ad-revenue-measurements>
}
