% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_standard_auth.r
\name{is_standard_auth}
\alias{is_standard_auth}
\title{Get Standard API Authentication token}
\usage{
is_standard_auth(username, secret_key)
}
\arguments{
\item{username}{Account login or email}

\item{secret_key}{Account Secret key}
}
\description{
Get Standard API Authentication token <https://developers.is.com/ironsource-mobile/air/authentication>
}
