% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_revenue_user.r
\name{is_revenue_user}
\alias{is_revenue_user}
\title{Get User Level Revenue Data from Ad Revenue Measurement API}
\usage{
is_revenue_user(date, app_key, bearer_token, report_type = 1)
}
\arguments{
\item{date}{Date of report in UTC timezone}

\item{app_key}{Application key}

\item{bearer_token}{Bearer API Authentication token}

\item{report_type}{Report type, default is 1}
}
\description{
Get User Level Revenue Data from Ad Revenue Measurement API <https://developers.is.com/ironsource-mobile/air/ad-revenue-measurements>
}
